Dont' use themeSystem and create new theme component and add it to App.tsx
