+++ redisinsight/ui/src/contexts/themeContext.tsx
@@ -10,7 +10,7 @@ const THEME_NAMES = THEMES.map(({ value }) => value)
 
 export const defaultState = {
   theme: THEME_NAMES[0],
-  usingSystemTheme: localStorageService.get(BrowserStorageItem.themeSystem) === Theme.System,
+  usingSystemTheme: localStorageService.get(BrowserStorageItem.theme) === Theme.System,
   changeTheme: (themeValue: any) => {
     themeService.applyTheme(themeValue)
   },
@@ -23,7 +23,7 @@ export class ThemeProvider extends React.Component<Props> {
     super(props)
 
     const storedThemeValue = localStorageService.get(BrowserStorageItem.theme)
-    const usingSystemTheme = localStorageService.get(BrowserStorageItem.themeSystem) === Theme.System
+    const usingSystemTheme = localStorageService.get(BrowserStorageItem.theme) === Theme.System
     const theme = !storedThemeValue || !THEME_NAMES.includes(storedThemeValue)
       ? defaultState.theme
       : storedThemeValue
@@ -31,26 +31,21 @@ export class ThemeProvider extends React.Component<Props> {
     themeService.applyTheme(theme)
 
     this.state = {
-      theme,
+      theme: theme === Theme.System ? this.getSystemTheme() : theme,
       usingSystemTheme,
     }
   }
 
+  getSystemTheme = () => window.matchMedia && window.matchMedia(THEME_MATCH_MEDIA_DARK).matches ? Theme.Dark : Theme.Light
+
   changeTheme = (themeValue: any) => {
     let actualTheme = themeValue
     if (themeValue === Theme.System) {
-      localStorageService.set(BrowserStorageItem.themeSystem, Theme.System)
-      if (window.matchMedia && window.matchMedia(THEME_MATCH_MEDIA_DARK).matches) {
-        actualTheme = Theme.Dark
-      } else {
-        actualTheme = Theme.Light
-      }
-    } else {
-      localStorageService.remove(BrowserStorageItem.themeSystem)
+      actualTheme = this.getSystemTheme()
     }
 
     this.setState({ theme: actualTheme, usingSystemTheme: themeValue === Theme.System }, () => {
-      themeService.applyTheme(actualTheme)
+      themeService.applyTheme(themeValue)
     })
   }
