+++ redisinsight/ui/src/services/theme.ts
@@ -1,5 +1,5 @@
 import { isFunction } from 'lodash'
-import { BrowserStorageItem, Theme } from 'uiSrc/constants'
+import { BrowserStorageItem, Theme, THEME_MATCH_MEDIA_DARK } from 'uiSrc/constants'
 import { localStorageService } from './storage'
 
 interface CSSFile {
@@ -19,6 +19,15 @@ class ThemeService {
   }
 
   applyTheme(newTheme: Theme) {
+    let actualTheme = newTheme
+    if (newTheme === Theme.System) {
+      if (window.matchMedia && window.matchMedia(THEME_MATCH_MEDIA_DARK).matches) {
+        newTheme = Theme.Dark
+      } else {
+        newTheme = Theme.Light
+      }
+    }
+
     Object.values(this.themes).forEach((theme: CSSFile[]) =>
       theme.forEach(
         (cssFile: CSSFile) => isFunction(cssFile?.unuse) && cssFile.unuse()
@@ -26,7 +35,7 @@ class ThemeService {
     this.themes[newTheme].forEach(
       (cssFile: CSSFile) => isFunction(cssFile?.use) && cssFile.use()
     )
-    localStorageService.set(BrowserStorageItem.theme, newTheme)
+    localStorageService.set(BrowserStorageItem.theme, actualTheme)
   }
 
   static getTheme() {
